requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
click==8.1.7
cmake==3.30.2
colorama==0.4.6
Django==5.0.7
django-filter==24.2
django-image-uploader-widget==0.5.0
django-jazzmin==3.0.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
dlib==19.24.99
environs==11.0.0
eskiz-sms==0.2.3
face-recognition==1.3.0
face_recognition_models==0.3.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.7
marshmallow==3.21.3
numpy==2.0.1
packaging==24.1
phonenumbers==8.13.40
pillow==10.4.0
python-dotenv==1.0.1
redis==5.0.8
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.1
